+++ game/settings/settings.py
@@ -329,6 +329,14 @@ class Settings:
         min=0,
         max=100,
     )
+    armed_recon_engagement_range_distance: int = bounded_int_option(
+        "Armed Recon engagement range (NM)",
+        page=CAMPAIGN_DOCTRINE_PAGE,
+        section=DOCTRINE_DISTANCES_SECTION,
+        default=5,
+        min=0,
+        max=25,
+    )
     sead_sweep_engagement_range_distance: int = bounded_int_option(
         "SEAD Sweep engagement range (NM)",
         page=CAMPAIGN_DOCTRINE_PAGE,